search_job/requirements.txt
fastapi
uvicorn
requests
python-multipart
PyPDF2
python-docx
